[utils] Add suppoort for negative arguments in Barret reduction

+++ codeforces/1992/g/main.cc
@@ -34,7 +34,7 @@ template <typename T, auto M> struct Mod {
   static V inv(V x, V m) { return x > 1 ? m - inv(m % x, x) * m / x : 1; }
   make_unsigned_t<T> x;
   Mod() : x(0) {}
-  Mod(auto y) : x(y % i64(M)) { x >= M ? x += M : x; }
+  Mod(auto y) : x(y % M) { x >= M ? x += M : x; }
   operator T() const { return x; }
   Mod operator-() const { return Mod() -= *this; }
   Mod operator+(auto rhs) const { return Mod(*this) += rhs; }
@@ -47,7 +47,7 @@ template <typename T, auto M> struct Mod {
   Mod &operator/=(Mod rhs) { return x = x * inv(rhs.x, M) % M, *this; }
   Mod pow(auto y) const { // O(log y) | 0^(-inf,0] -> 1
     Mod ans(1), base(*this);
-    for (auto e = y < 0 ? ~y + 1ull : +y; e; e >>= 1, base *= base) {
+    for (auto e = y < 0 ? ~y + u128(1) : +y; e; e >>= 1, base *= base) {
       e & 1 ? ans *= base : ans;
     }
     return y < 0 ? Mod(1) /= ans : ans;
